+++ dao-contracts/bin/mocks/mock_owner_contract.rs
@@ -1,0 +1,4 @@
+use casper_dao_contracts::mocks::{MockOwner, MockOwnerInterface};
+casper_dao_contracts::mock_owner!();
+
+fn main() {}
