+++ dao-contracts/bin/mocks/mock_repository_contract.rs
@@ -1,0 +1,4 @@
+use casper_dao_contracts::mocks::{MockRepository, MockRepositoryInterface};
+casper_dao_contracts::mock_repository!();
+
+fn main() {}
