+++ dao-contracts/bin/mocks/mock_staking_contract.rs
@@ -1,0 +1,4 @@
+use casper_dao_contracts::mocks::{MockStaking, MockStakingInterface};
+casper_dao_contracts::mock_staking!();
+
+fn main() {}
