+++ dao-contracts/bin/mocks/mock_token_contract.rs
@@ -1,0 +1,4 @@
+use casper_dao_contracts::mocks::{MockToken, MockTokenInterface};
+casper_dao_contracts::mock_token!();
+
+fn main() {}
