+++ dao-contracts/bin/mocks/mock_voter_contract.rs
@@ -1,4 +1,4 @@
-use casper_dao_contracts::mocks::mock_voter::{MockVoterContract, MockVoterContractInterface};
+use casper_dao_contracts::mocks::{MockVoterContract, MockVoterContractInterface};
 casper_dao_contracts::mock_voter_contract!();
 
 fn main() {}
