+++ dao-contracts/src/lib.rs
@@ -4,9 +4,6 @@ mod reputation;
 mod variable_repository;
 pub mod voting;
 
-pub use mocks::mock_voter::{
-    MockVoterContract, MockVoterContractCaller, MockVoterContractInterface,
-};
 pub use repo_voter::{RepoVoterContract, RepoVoterContractCaller, RepoVoterContractInterface};
 pub use reputation::{ReputationContract, ReputationContractCaller, ReputationContractInterface};
 pub use variable_repository::{
