+++ dao-contracts/src/mocks/mock_whitelist.rs
@@ -1,3 +1,4 @@
+use casper_dao_modules::Whitelist;
 use casper_dao_utils::{
     casper_dao_macros::{casper_contract_interface, Instance},
     Address,
@@ -5,13 +6,11 @@ use casper_dao_utils::{
 
 use delegate::delegate;
 
-use crate::Whitelist;
-
 #[casper_contract_interface]
 trait MockWhitelistInterface {
     fn init(&mut self) {}
-    fn add(&mut self, address: Address);
-    fn remove(&mut self, address: Address);
+    fn add_to_whitelist(&mut self, address: Address);
+    fn remove_from_whitelist(&mut self, address: Address);
     fn ensure_whitelisted(&self);
     fn is_whitelisted(&self, address: Address) -> bool;
 }
@@ -24,10 +23,8 @@ pub struct MockWhitelist {
 impl MockWhitelistInterface for MockWhitelist {
     delegate! {
         to self.whitelist {
-            #[call(add_to_whitelist)]
-            fn add(&mut self, address: Address);
-            #[call(remove_from_whitelist)]
-            fn remove(&mut self, address: Address);
+            fn add_to_whitelist(&mut self, address: Address);
+            fn remove_from_whitelist(&mut self, address: Address);
             fn ensure_whitelisted(&self);
         }
     }
