+++ dao-modules/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod mocks;
 mod owner;
 mod repository;
 mod staking;
