--- dao-modules/src/mocks/mod.rs
@@ -1,15 +1,0 @@
-mod mock_owner;
-mod mock_repository;
-mod mock_staking;
-mod mock_token;
-mod mock_whitelist;
-
-pub use mock_owner::{MockOwner, MockOwnerCaller, MockOwnerInterface, MockOwnerTest};
-pub use mock_repository::{
-    MockRepository, MockRepositoryCaller, MockRepositoryInterface, MockRepositoryTest,
-};
-pub use mock_staking::{MockStaking, MockStakingCaller, MockStakingInterface, MockStakingTest};
-pub use mock_token::{MockToken, MockTokenCaller, MockTokenInterface, MockTokenTest};
-pub use mock_whitelist::{
-    MockWhitelist, MockWhitelistCaller, MockWhitelistInterface, MockWhitelistTest,
-};
